embedding-tool & text2SQL/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
openai==1.12.0
google-generativeai>=0.7.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
